% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{bootstrap for finding confidence interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{vector}

\item{fun}{function}

\item{alpha}{alpha}

\item{cx}{cx}
}
\value{
graph
}
\description{
bootstrap for finding confidence interval
}
\examples{
myboot2(iter = 10000, fun = "mean", alpha = 0.05, cx = 1.5)
}
