% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myctl.R
\name{myclt}
\alias{myclt}
\title{A function for graphing samples}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{number of samples}

\item{iter}{number of iterations}
}
\value{
a numeric vector
}
\description{
A function for graphing samples
}
\examples{
myctl(3,21)

}
