% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot a Normal Distribution Curve and Calculate Probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{Upper limit of integration for calculating the probability.}
}
\value{
A list containing the mean (mu), standard deviation (sigma), and
probability (P(X <= a)).
}
\description{
Plot a Normal Distribution Curve and Calculate Probability
}
\examples{
myncurve(0, 1, 1.96)
}
