% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYNRML.R
\name{myNRML}
\alias{myNRML}
\title{A quadratic function}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
A quadratic function
}
\examples{
myquad(x = 1:10 )

}
