% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{A sampling function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{x}{3 integers}
}
\value{
barplot and numeric value
}
\description{
A sampling function
}
\examples{
mysample(10, iter = 10, time = 0.5)
}
